chore: helper file for common methods
