+++ src/session-pool.ts
@@ -31,6 +31,14 @@ import {
   startTrace,
 } from './instrument';
 
+import {
+  isDatabaseNotFoundError,
+  isInstanceNotFoundError,
+  isDefaultCredentialsNotSetError,
+  isProjectIdNotSetInEnvironmentError,
+  isCreateSessionPermissionError,
+} from './helper';
+
 /**
  * @callback SessionPoolCloseCallback
  * @param {?Error} error Closing error, if any.
@@ -238,81 +246,6 @@ export function isSessionNotFoundError(
     error !== undefined &&
     error.code === grpc.status.NOT_FOUND &&
     error.message.includes('Session not found')
-  );
-}
-
-/**
- * Checks whether the given error is a 'Database not found' error.
- * @param {Error} error The error to check.
- * @return {boolean} True if the error is a 'Database not found' error, and otherwise false.
- */
-export function isDatabaseNotFoundError(
-  error: grpc.ServiceError | undefined
-): boolean {
-  return (
-    error !== undefined &&
-    error.code === grpc.status.NOT_FOUND &&
-    error.message.includes('Database not found')
-  );
-}
-
-/**
- * Checks whether the given error is an 'Instance not found' error.
- * @param {Error} error The error to check.
- * @return {boolean} True if the error is an 'Instance not found' error, and otherwise false.
- */
-export function isInstanceNotFoundError(
-  error: grpc.ServiceError | undefined
-): boolean {
-  return (
-    error !== undefined &&
-    error.code === grpc.status.NOT_FOUND &&
-    error.message.includes('Instance not found')
-  );
-}
-
-/**
- * Checks whether the given error is a 'Create session permission' error.
- * @param {Error} error The error to check.
- * @return {boolean} True if the error is a 'Create session permission' error, and otherwise false.
- */
-export function isCreateSessionPermissionError(
-  error: grpc.ServiceError | undefined
-): boolean {
-  return (
-    error !== undefined &&
-    error.code === grpc.status.PERMISSION_DENIED &&
-    error.message.includes('spanner.sessions.create')
-  );
-}
-
-/**
- * Checks whether the given error is a 'Could not load the default credentials' error.
- * @param {Error} error The error to check.
- * @return {boolean} True if the error is a 'Could not load the default credentials' error, and otherwise false.
- */
-export function isDefaultCredentialsNotSetError(
-  error: grpc.ServiceError | undefined
-): boolean {
-  return (
-    error !== undefined &&
-    error.message.includes('Could not load the default credentials')
-  );
-}
-
-/**
- * Checks whether the given error is an 'Unable to detect a Project Id in the current environment' error.
- * @param {Error} error The error to check.
- * @return {boolean} True if the error is an 'Unable to detect a Project Id in the current environment' error, and otherwise false.
- */
-export function isProjectIdNotSetInEnvironmentError(
-  error: grpc.ServiceError | undefined
-): boolean {
-  return (
-    error !== undefined &&
-    error.message.includes(
-      'Unable to detect a Project Id in the current environment'
-    )
   );
 }
